+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros/loader.py
@@ -30,7 +30,7 @@ from mosaicolabs.logging_config import get_logger
 from mosaicolabs.models.core.helpers import resolve_ontology_class
 
 from ...models.core.serializable import _compute_schema_fingerprint
-from ..protocols.mcaps.converters.ros_converter import RosMsgConverter
+from ..protocols.mcap.converters.ros_converter import RosMsgConverter
 from .adapter_base import ROSAdapterBase, RosSchemaMetadata
 from .helpers import (
     _class_name_from_ros_msgtype,
